chore: move all v2 endpoints to v1
