+++ bc_obps/registration/api/__init__.py
@@ -1,3 +1,45 @@
 # ruff: noqa: F403
-from .v2 import *
+from . import (
+    business_structures,
+    contacts,
+    e2e_test_setup,
+    facilities,
+    naics_codes,
+    operations,
+    operators,
+    regulated_products,
+    reporting_activities,
+    user_operators,
+    users,
+)
+from ._operations import current as operations_current
+from ._operations._operation_id import boro_id, bcghg_id as operation_bcghg_id, operation_representatives
+from ._operations._operation_id.facilities import list_facilities_by_operation_id
+from ._facilities._facility_id import bcghg_id as facility_bcghg_id
+from ._facilities import facility_id
+from ._operators import search, operator_id
+from .transfer_events import list_transfer_events
+from ._operators._operator_id import has_admin, request_access, request_admin_access, access_declined, confirm
+from ._contacts import contact_id
+from ._operations import operation_id
+from ._operations._operation_id._registration import (
+    new_entrant_application,
+    operation,
+    operation_representative,
+    opted_in_operation,
+    submission,
+)
+from ._user_operators import current, pending, user_operator_id
+from ._user_operators._current import (
+    access_requests,
+    has_registered_operation,
+    has_required_fields,
+    is_current_user_approved_admin,
+    operator,
+    operator_users,
+)
+from ._user_operators._user_operator_id import update_status
+from .user import user_profile, user_app_role
+from . import transfer_events
+from ._transfer_events import transfer_id
 from .router import router  # django_ninja router object  # noqa: F401
