+++ bc_obps/registration/api/_operations/_operation_id/_registration/operation.py
