+++ bc_obps/registration/api/_operators/_operator_id/confirm.py
