+++ bc_obps/registration/api/_user_operators/_current/has_registered_operation.py
