+++ bc_obps/registration/api/_user_operators/_current/has_required_fields.py
