+++ bc_obps/registration/api/facilities.py
@@ -6,7 +6,7 @@ from registration.models.facility import Facility
 from registration.schema import FacilityIn, FacilityOut, Message
 from service.facility_service import FacilityService
 from common.api.utils import get_current_user_guid
-from ..router import router
+from registration.api.router import router
 from service.error_service.custom_codes_4xx import custom_codes_4xx
