+++ bc_obps/registration/api/regulated_products.py
@@ -2,7 +2,7 @@ from django.http import HttpRequest
 from registration.constants import REGULATED_PRODUCT_TAGS
 from service.data_access_service.regulated_product_service import RegulatedProductDataAccessService
 from common.permissions import authorize
-from ..router import router
+from registration.api.router import router
 from typing import List, Literal, Tuple
 from registration.models import RegulatedProduct
 from registration.schema import RegulatedProductSchema
