+++ bc_obps/registration/schema/validators.py
@@ -1,9 +1,5 @@
 from typing import Optional
 
-from registration.utils import data_url_to_file
-from django.core.files.base import ContentFile
-
-
 def validate_cra_business_number(value: int) -> Optional[int]:
     """
     Validate that the cra_business_number has 9 digits.
